+++ clx_task_management/__manifest__.py
@@ -3,7 +3,7 @@
 # See LICENSE file for full copyright & licensing details.
 {
     'name': 'CLX Task Management',
-    'version': '13.3.0.0.4',
+    'version': '13.3.0.0.5',
     'summary': 'CLX Task Management',
     'sequence': 1,
     'description': """ CLX Task Management """,
@@ -21,7 +21,9 @@
         'views/team_views.xml',
         'views/sub_task_views.xml',
         'views/request_form_views.xml',
-        'views/project_views.xml'
+        'views/project_views.xml',
+        'views/sale_order_views.xml',
+        'wizard/task_cancel_warning_wizard_views.xml'
     ],
     'demo': [
     ],