+++ clx_task_management/models/__init__.py
@@ -6,3 +6,5 @@ from . import sub_task
 from . import team
 from . import request_form
 from . import project
+from . import sale_order_line
+from . import sale_order