+++ clx_task_management/models/sale_order_line.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo, CLx Media
+# See LICENSE file for full copyright & licensing details.
+from odoo import fields, models
+
+
+class SaleOrderLine(models.Model):
+    _inherit = 'sale.order.line'
+
+    project_id = fields.Many2one('project.project', string="Project")
+    task_id = fields.Many2one('project.task', string='Task')