+++ clx_task_management/wizard/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo, CLx Media
 # See LICENSE file for full copyright & licensing details.
-from . import models
-from . import wizard
+from . import task_cancel_warning_wizard